[build] rename storage packages;
