+++ internal/driven/storage/battlestrg/storage.go
@@ -1,4 +1,4 @@
-package battle
+package battlestrg
 
 import (
 	"context"
