+++ internal/driven/storage/battlestrg/storage_test.go
@@ -1,4 +1,4 @@
-package battle
+package battlestrg
 
 import "testing"
