+++ internal/driven/storage/gamestrg/storage.go
@@ -1,4 +1,4 @@
-package game
+package gamestrg
 
 import (
 	"context"
