+++ internal/driven/storage/gamestrg/storage_test.go
@@ -1,4 +1,4 @@
-package game
+package gamestrg
 
 import "testing"
