+++ internal/driven/storage/pokestrg/storage_test.go
@@ -1,4 +1,4 @@
-package pokemon
+package pokestrg
 
 import "testing"
